Delete dialog style added

+++ src/app/user-details/user-details.component.ts
@@ -14,6 +14,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 export class UserDetailsComponent implements OnInit {
   user$ : Observable<any>;
   invalidDelete;
+  noDialog = true;
   constructor(private usersService: UsersService, private route: ActivatedRoute, private router: Router, 
     private deleteUserService: DeleteUserService, private dialog: MatDialog) {
      }
@@ -36,15 +37,19 @@ export class UserDetailsComponent implements OnInit {
   }
 
   openDialog(firstName){
-    this.dialog.open(ConfirmDialogComponent, {
-      data: { name: firstName },
-    }).afterClosed().subscribe(
-      result => {
-        if(result){
-          this.delete();
+    if(this.noDialog){
+      this.dialog.open(ConfirmDialogComponent, {
+        data: { name: firstName }
+      }).afterClosed().subscribe(
+        result => {
+          this.noDialog = true;
+          if(result){
+            this.delete();
+          }
         }
-      }
-    );
+      );
+    }
+    this.noDialog = false;
   }
